# .sisyphus/drafts/tmuxweb-openapi.yaml
openapi: 3.0.3
info:
  title: TmuxWeb API
  description: |
    API for TmuxWeb - a web-based tmux session manager with task tracking,
    conversation logging, and summary generation capabilities.
    
    ## Authentication
    TmuxWeb uses cookie-based session authentication:
    - Login with a token via `POST /api/auth/login`
    - Session cookie is HttpOnly, Max-Age=30 days, SameSite=Lax, Secure in production
    - All authenticated endpoints require a valid session cookie
    
    ## paneKey Format
    The paneKey identifies a specific tmux pane: `"{sessionName}:{windowIndex}:{paneIndex}"`
    
    Example: `"hsk:1:0"` = session "hsk", window 1, pane 0
  version: 1.0.0
  contact:
    name: TmuxWeb
servers:
  - url: http://localhost:8215
    description: Local development server
...
paths:
...
  /api/profiles:
    get:
      tags: [Profiles]
      summary: List profiles for current token
      description: Returns all profiles (workspaces) for the authenticated user
      operationId: listProfiles
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileListResponse'
              example:
                profiles:
                  - id: 1
                    profile_key: "default"
                    name: "Default"
                    ctime: 1707400000
                    mtime: 1707400000
                  - id: 2
                    profile_key: "work"
                    name: "Work"
                    ctime: 1707400100
                    mtime: 1707400100
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Profiles]
      summary: Create a new profile
      operationId: createProfile
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
            example:
              name: "Personal"
              profile_key: "personal"
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
              example:
                id: 3
                profile_key: "personal"
                name: "Personal"
                ctime: 1707400200
                mtime: 1707400200
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/profiles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: Profile ID

    put:
      tags: [Profiles]
      summary: Update a profile
      operationId: updateProfile
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              name: "Personal Projects"
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Profiles]
      summary: Delete a profile
      operationId: deleteProfile
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Profile deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
              example:
                success: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # Session Groups
  # ============================================================================
...
  /api/profiles/{id}/order:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: Profile ID

    get:
      tags: [Ordering]
      summary: Get current order for profile
      description: Returns ordered list of groups and sessions for the profile
      operationId: getOrder
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                order:
                  - type: "session"
                    session_name: "dify"
                    sort_order: 10
                  - type: "group"
                    group_id: 1
                    group_name: "AI Projects"
                    sort_order: 10
                    sessions:
                      - session_name: "hsk"
                        sort_order: 10
                      - session_name: "omo"
                        sort_order: 20
                  - type: "session"
                    session_name: "misc"
                    sort_order: 20
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Ordering]
      summary: Save new order for profile
      description: Updates sort_order for groups and sessions
      operationId: updateOrder
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
            example:
              order:
                - type: "group"
                  group_id: 1
                  sort_order: 5
                  sessions:
                    - session_name: "omo"
                      sort_order: 10
                    - session_name: "hsk"
                      sort_order: 20
                - type: "session"
                  session_name: "dify"
                  sort_order: 10
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
              example:
                success: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # Pane Status
  # ============================================================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        HttpOnly session cookie set by /api/auth/login.
        Cookie settings: Max-Age=30 days, SameSite=Lax, Secure in production.
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
            message: "Authentication required"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "bad_request"
            message: "Invalid input"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not_found"
            message: "Resource not found"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Optional error code

    # --------------------------------------------------------------------------
    # Auth
    # --------------------------------------------------------------------------
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        profile_key:
          type: string
          description: Unique profile identifier (UUID or slug)
        name:
          type: string
        ctime:
          type: integer
          description: Creation timestamp
        mtime:
          type: integer
          description: Modification timestamp
    ProfileListResponse:
      type: object
      properties:
        profiles:
          type: array
          items:
            $ref: '#/components/schemas/Profile'
    CreateProfileRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        profile_key:
          type: string
          description: Optional custom key (auto-generated if not provided)
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string

    # --------------------------------------------------------------------------
    # Groups
    # --------------------------------------------------------------------------
...
    OrderItem:
      type: object
      properties:
        type:
          type: string
          enum: [session, group]
        session_name:
          type: string
          description: Present when type=session
        group_id:
          type: integer
          description: Present when type=group
        group_name:
          type: string
          description: Present when type=group
        sort_order:
          type: integer
        sessions:
          type: array
          items:
            type: object
            properties:
              session_name:
                type: string
              sort_order:
                type: integer
          description: Present when type=group
    OrderResponse:
      type: object
      properties:
        order:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    UpdateOrderRequest:
      type: object
      required:
        - order
      properties:
        order:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    # --------------------------------------------------------------------------
    # Pane Status
    # --------------------------------------------------------------------------
...